refactor : DI 적용을 위한 Dao Interface 추가로 인한 수정
