+++ src/main/java/core/mvc/LegacyRequestMapping.java
@@ -3,9 +3,9 @@ package core.mvc;
 import core.nmvc.HandlerMapping;
 import next.controller.HomeController;
 import next.controller.qna.*;
-import next.dao.AnswerDao;
-import next.dao.QuestionDao;
-import next.dao.UserDao;
+import next.dao.impl.JdbcAnswerDao;
+import next.dao.impl.JdbcQuestionDao;
+import next.dao.impl.JdbcUserDao;
 import next.service.qna.QuestionService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -43,9 +43,9 @@ public class LegacyRequestMapping implements HandlerMapping {
     /*
      * dao
      */
-    var userDao = new UserDao();
-    var questionDao = new QuestionDao();
-    var answerDao = new AnswerDao();
+    var userDao = new JdbcUserDao();
+    var questionDao = new JdbcQuestionDao();
+    var answerDao = new JdbcAnswerDao();
 
     /*
      * service
