+++ src/main/java/next/controller/HomeController.java
@@ -2,7 +2,7 @@ package next.controller;
 
 import core.mvc.AbstractController;
 import core.mvc.view.ModelAndView;
-import next.dao.QuestionDao;
+import next.dao.impl.JdbcQuestionDao;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -10,15 +10,15 @@ import java.io.IOException;
 
 public class HomeController extends AbstractController {
 
-  private final QuestionDao questionDao;
+  private final JdbcQuestionDao jdbcQuestionDao;
 
-  public HomeController(QuestionDao questionDao) {
-    this.questionDao = questionDao;
+  public HomeController(JdbcQuestionDao jdbcQuestionDao) {
+    this.jdbcQuestionDao = jdbcQuestionDao;
   }
 
   @Override
   public ModelAndView execute(HttpServletRequest request, HttpServletResponse response)
       throws IOException {
-    return jspView("/home.jsp").addObject("questions", questionDao.findAll());
+    return jspView("/home.jsp").addObject("questions", jdbcQuestionDao.findAll());
   }
 }
