+++ src/main/java/next/controller/qna/AddQuestionController.java
@@ -4,6 +4,7 @@ import core.mvc.AbstractController;
 import core.mvc.view.ModelAndView;
 import next.controller.UserSessionUtils;
 import next.dao.QuestionDao;
+import next.dao.impl.JdbcQuestionDao;
 import next.model.Question;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
