+++ src/main/java/next/controller/qna/RemoveAnswerController.java
@@ -4,6 +4,8 @@ import core.mvc.AbstractController;
 import core.mvc.view.ModelAndView;
 import next.dao.AnswerDao;
 import next.dao.QuestionDao;
+import next.dao.impl.JdbcAnswerDao;
+import next.dao.impl.JdbcQuestionDao;
 import next.model.Result;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
