+++ src/main/java/next/controller/qna/RemoveQuestionController.java
@@ -2,18 +2,12 @@ package next.controller.qna;
 
 import core.mvc.AbstractController;
 import core.mvc.view.ModelAndView;
-import next.controller.UserSessionUtils;
-import next.dao.QuestionDao;
-import next.dao.UserDao;
 import next.service.qna.QuestionService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-import static org.apache.commons.lang3.ObjectUtils.isEmpty;
 import static org.apache.commons.lang3.math.NumberUtils.toLong;
 
 public class RemoveQuestionController extends AbstractController {
