+++ src/main/java/next/controller/user/UserController.java
@@ -7,6 +7,7 @@ import core.mvc.view.ModelAndView;
 import core.nmvc.AbstractNewController;
 import next.controller.UserSessionUtils;
 import next.dao.UserDao;
+import next.dao.impl.JdbcUserDao;
 import next.model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -20,11 +21,11 @@ public class UserController extends AbstractNewController {
 
   private final Logger logger = LoggerFactory.getLogger(getClass());
 
-  private final UserDao userDao = new UserDao();
+  private final UserDao userDao = new JdbcUserDao();
 
-//  public UserController(UserDao userDao) {
-//    this.userDao = userDao;
-//  }
+  //  public UserController(UserDao userDao) {
+  //    this.userDao = userDao;
+  //  }
 
   @RequestMapping("/users/list")
   public ModelAndView list(HttpServletRequest request, HttpServletResponse response)
