+++ src/main/java/next/dao/UserDao.java
@@ -1,47 +1,16 @@
 package next.dao;
 
-import core.jdbc.JdbcTemplate;
-import core.jdbc.RowMapper;
 import next.model.User;
 
 import java.util.List;
 
-public class UserDao {
+public interface UserDao {
 
-  private final JdbcTemplate template = JdbcTemplate.getInstance();
+    void insert(User user);
 
-  private static final RowMapper<User> MAPPER =
-      rs ->
-          new User(
-              rs.getString("userId"),
-              rs.getString("password"),
-              rs.getString("name"),
-              rs.getString("email"));
+    void update(User user);
 
-  public void insert(User user) {
-    var sql = "INSERT INTO USERS VALUES (?, ?, ?, ?)";
+    List<User> findAll();
 
-    template.update(sql, user.getUserId(), user.getPassword(), user.getName(), user.getEmail());
-  }
-
-  public void update(User user) {
-
-    var sql = "UPDATE USERS SET password=?, name=?, email=? WHERE userId=?";
-
-    template.update(sql, user.getPassword(), user.getName(), user.getEmail(), user.getUserId());
-  }
-
-  public List<User> findAll() {
-
-    var sql = "SELECT userId, password, name, email FROM USERS";
-
-    return template.query(sql, null, MAPPER);
-  }
-
-  public User findByUserId(String userId) {
-
-    String sql = "SELECT userId, password, name, email FROM USERS WHERE userid=?";
-
-    return template.queryForObject(sql, new Object[] {userId}, MAPPER);
-  }
+    User findByUserId(String userId);
 }
