+++ src/main/java/next/service/qna/QuestionService.java
@@ -1,8 +1,8 @@
 package next.service.qna;
 
 import next.controller.UserSessionUtils;
-import next.dao.QuestionDao;
-import next.dao.UserDao;
+import next.dao.impl.JdbcQuestionDao;
+import next.dao.impl.JdbcUserDao;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -14,31 +14,31 @@ public class QuestionService {
 
   private final Logger logger = LoggerFactory.getLogger(getClass());
 
-  private final UserDao userDao;
-  private final QuestionDao questionDao;
+  private final JdbcUserDao jdbcUserDao;
+  private final JdbcQuestionDao jdbcQuestionDao;
 
-  public QuestionService(UserDao userDao, QuestionDao questionDao) {
-    this.userDao = userDao;
-    this.questionDao = questionDao;
+  public QuestionService(JdbcUserDao jdbcUserDao, JdbcQuestionDao jdbcQuestionDao) {
+    this.jdbcUserDao = jdbcUserDao;
+    this.jdbcQuestionDao = jdbcQuestionDao;
   }
 
   public void removeQuestion(HttpSession session, long questionId) {
 
-    var question = questionDao.findById(questionId);
+    var question = jdbcQuestionDao.findById(questionId);
 
     if (isEmpty(question)) {
       throw new IllegalStateException("no exist question.");
     }
 
-    if (!UserSessionUtils.isSameUser(session, userDao.findByUserId(question.getWriter()))) {
+    if (!UserSessionUtils.isSameUser(session, jdbcUserDao.findByUserId(question.getWriter()))) {
       throw new IllegalStateException("No same question user.");
     }
 
-    if (!questionDao.checkDelete(question.getQuestionId(), question.getWriter())) {
+    if (!jdbcQuestionDao.checkDelete(question.getQuestionId(), question.getWriter())) {
       throw new IllegalStateException("can not remove question.");
     }
 
-    questionDao.delete(question.getQuestionId());
+    jdbcQuestionDao.delete(question.getQuestionId());
 
     logger.debug("delete questionId : {}", question.getQuestionId());
   }
